# Conflicts:
#	README.MD

+++ src/main/java/eu/tjago/apps/ivonatalker/IvonaTalkerApp.java
@@ -16,17 +16,18 @@ import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.BorderPane;
+import javafx.stage.FileChooser;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
 
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.Marshaller;
-import javax.xml.bind.Unmarshaller;
 import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.util.prefs.Preferences;
+import java.nio.file.Path;
+import java.nio.file.StandardCopyOption;
+import java.util.Optional;
 
 /**
  * Created by Tomasz on 2015-12-24.
@@ -64,8 +65,14 @@ public class IvonaTalkerApp extends Application {
     @Override
     public void stop() throws Exception {
 
+        //FIXME release player resource file
+
         // remove temporary .mp3 file
-        Files.delete(SpeechCloudSingleton.getTmpSpeechFile().toPath());
+        Path path = SpeechCloudSingleton.getTmpSpeechFile().toPath();
+
+        if (Files.exists(path)) {
+            Files.delete(SpeechCloudSingleton.getTmpSpeechFile().toPath());
+        }
 
         Platform.exit();
     }
@@ -214,6 +221,40 @@ public class IvonaTalkerApp extends Application {
             e.printStackTrace();
 
             alert.showAndWait();
+        }
+    }
+
+    public void saveVoiceToFile() {
+        FileChooser fileChooser = new FileChooser();
+
+        // Set extension filter
+        FileChooser.ExtensionFilter extFilter = new FileChooser.ExtensionFilter("MP3 files (*.mp3)", "*.mp3");
+        fileChooser.getExtensionFilters().add(extFilter);
+
+        //set init Dir upon opening Dialog
+        Optional<File> initialPath = Optional.ofNullable(FileHelper.getLastSavedLocation());
+
+        if (initialPath.isPresent()) {
+            fileChooser.setInitialDirectory(initialPath.get());
+        }
+
+        // Show save file dialog
+        File file = fileChooser.showSaveDialog(this.primaryStage);
+
+        if (file != null) {
+            // Make sure it has the correct extension
+            if (!file.getPath().endsWith(".mp3")) {
+                file = new File(file.getPath() + ".mp3");
+            }
+            try {
+                Files.copy(SpeechCloudSingleton.getTmpSpeechFile().toPath(),
+                        file.toPath(),
+                        StandardCopyOption.REPLACE_EXISTING);
+                FileHelper.setLastSavedLocation(file);
+
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
     }
